Split blocks grouping and item to prevent nested styling in PagePartial
