+++ src/components/Blocks/PagePartial/PagePartial.vue
@@ -1,11 +1,12 @@
 <template>
-  <Blocks v-if="section" :blocks="[section]" />
+  <BlockItem v-if="section" :block="section" />
 </template>
 
 <script setup lang="ts">
 import { type PagePartialFragment } from "./PagePartial.query";
 import { type FragmentOf, readFragment } from "~/utils/graphql";
-import type { BlockItem } from "./types";
+import type { BlockRecord } from "../types";
+import BlockItem from "../BlockItem.vue";
 
 const props = defineProps<{
   data: FragmentOf<typeof PagePartialFragment>;
@@ -14,7 +15,7 @@ const props = defineProps<{
 const data = readFragment<typeof PagePartialFragment>(props.data);
 
 type SupportedBlockItem = Exclude<
-  BlockItem,
+  BlockRecord,
   { __typename: "PagePartialBlockRecord" }
 >;
